Homework 4/services/analysis/fundamental/requirements.txt
setuptools
aiohttp
asyncpg
textblob
python-dotenv
